Remove flat_set
